Added Adjustable Ramp Intake
The ramp intake has 100hp, weighs 0.7t, and has a hitbox of 1.3mx1.9m, while the circular intake has 300hp and a hitbox of 1.3x0.7, less than half the size of the ramp intake. This makes the circular intake a far better pick for combat with a smaller hitbox and 3x the HP.

I added the ramp intake to this config with the same number of hitpoints as the ramp intake (300) to make it more viable. This makes the intakes a tradeoff between a 2x larger hitbox and more weight, or a 1/2 hitbox with less weight.