Odd-shaped part hitpoint balance
Patch to properly scale triangular wings to have half the hitpoints of equivalent square/rectangular wings, plus a couple additional nerfs/buffs to extreme outliers.

BDA calculates hitpoints of a part based off of its mass vs how large a 3d cube it fills. Because of this it does a poor job of scaling parts that are not rectangular. Ex. triangular wings have half the mass of square wings of the same dimensions, so the equation thinks the part is half as dense and scales its hitpoints accordingly: https://cdn.discordapp.com/attachments/712570695497809950/712627796731756594/unknown.png

 This patch balances a majority of non-rectangular parts with outlying hitpoints, and balances triangular wings to have half the hitpoints of equivalent square/rectangular wings. Each part has the in-game title and buff/nerf and reasoning next to it.